% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{dati}
\alias{dati}
\title{Data frame of meteorological data}
\format{A data frame with 800 rows and 4 variables:
\describe{
  \item{phen}{Temperature, Rain}
  \item{time}{ordered numbers for time}
  \item{zone}{label classification for group data}
  \item{value}{numbers for values}

}}
\usage{
dati
}
\description{
A dataset containing values to test outlier detection on meteorological number completely random
}
\keyword{datasets}
