% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\docType{data}
\name{ts}
\alias{ts}
\title{Time series}
\format{A data frame with 35 rows and 1 variable:
\describe{
  \item{dati}{pseudo random numers}

}}
\usage{
ts
}
\description{
A dataset containing a random time series with increasing trend and some variability
}
\keyword{datasets}
