% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{dati}
\alias{dati}
\title{Data frame of meteorological data}
\format{
A data frame with 800 rows and 4 variables:
\describe{
  \item{phen}{Temperature, Rain}
  \item{time}{ordered numbers for time (a number in the format YYYYMMDD [Year Month Day] is possible too)  }
  \item{zone}{label classification for the group, for example the identification code of a wheather station.}
  \item{value}{values}

}
}
\usage{
dati
}
\description{
This sample data would be invented meteorological information detected by weather stations.
}
\keyword{datasets}
